refactor(ai-engine): resolve #1745 — split recipe/custom_types.py (40K) into focused type modules
The 1282-line custom_types.py monolith was a single CustomTypesConverter
class holding Farmer's Delight + Create (17 convert_* methods, lines 65-1054)
and ImmersiveEngineering (4 convert_* methods, lines 1079-1266) conversion
logic plus the Forge recipe-type registry loader. Split into a mixin-based
composition that preserves the exact public API:

- converter_base.py (101 lines): CustomConverterBase (__init__ +
  _create_manual_review_result), _load_custom_recipe_types,
  CUSTOM_RECIPE_TYPES, is_custom_recipe_type.
- create_converter.py (1016 lines): CreateConverterMixin — Farmer's Delight
  (cooking_pot, cutting_board) + Create (mechanical_crafting, pressing,
  milling, crushing, deploying, splashing, compacting, mixing,
  sequenced_assembly, filling, emptying, cutting, haunting,
  sandpaper_polishing, item_application) + _build_create_conversion_note /
  _assemble_create_recipe helpers.
- immersive_engineering_converter.py (215 lines): ImmersiveEngineeringConverterMixin
  — crusher, metalpress, arc_furnace, refinery + _build_ie_* helpers.
- custom_types.py (65 lines): thin backwards-compat shim that composes
  CustomTypesConverter(CreateConverterMixin, ImmersiveEngineeringConverterMixin)
  and re-exports CUSTOM_RECIPE_TYPES / is_custom_recipe_type /
  _load_custom_recipe_types.

No public symbol renamed or removed; recipe/__init__.py imports resolve
unchanged via the shim (its own monolith refactor is tracked separately in
#1740, wave 11). Verified: 136 tests pass, ruff clean on new files.

Closes #1745.
